requirements.txt
numpy
pandas
scikit-learn
transformers
tqdm
tensorflow
sentence_transformers
faiss-cpu